Compute hashes & return JSON array in query()
Files are now indexed by their hash instead of their filename.  The
indices don't persist when the program is closed so the program might
take a while to start up if the shared folder contains a lot of data.

Instead of returning a set of files, LocalShares.query() now returns a
JSON array of search results, where a search result consists of a
file's name, size and hash.